refactor: remove reduntant tests
