--- packages/sdk/src/extensions/integrations/callOnIntegration/idleV4.test.ts
@@ -1,146 +1,0 @@
-import { IDLE_V4_ADAPTER } from "../../../../tests/constants.js";
-import { CLAIM_REWARDS_SELECTOR, LEND_SELECTOR, REDEEM_SELECTOR } from "../../../constants/selectors.js";
-import { decodeCallOnIntegrationArgs, encodeCallOnIntegrationArgs } from "./callOnIntegration.js";
-import { type Hex, getAddress, toHex } from "viem";
-import { expect, test } from "vitest";
-
-test("encodeCallOnIntegrationArgs for Idle V4 lend should encode correctly", () => {
-  expect(
-    encodeCallOnIntegrationArgs({
-      adapter: IDLE_V4_ADAPTER,
-      selector: LEND_SELECTOR,
-      integrationData: "0x1230",
-    }),
-  ).toMatchInlineSnapshot(
-    '"0x000000000000000000000000474c8d4a0e53b7235c6f8fc27c9b6406a32dd0b1099f751500000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000021230000000000000000000000000000000000000000000000000000000000000"',
-  );
-});
-
-test("decodeCallOnIntegrationArgs for Idle V4 lend should decode correctly", () => {
-  expect(
-    decodeCallOnIntegrationArgs(
-      "0x000000000000000000000000474c8d4a0e53b7235c6f8fc27c9b6406a32dd0b1099f751500000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000021230000000000000000000000000000000000000000000000000000000000000",
-    ),
-  ).toMatchInlineSnapshot(
-    {
-      adapter: IDLE_V4_ADAPTER,
-      selector: LEND_SELECTOR,
-      integrationData: "0x1230",
-    },
-    `
-    {
-      "adapter": "0x474c8d4A0e53B7235C6f8fC27c9B6406a32Dd0b1",
-      "integrationData": "0x1230",
-      "selector": "0x099f7515",
-    }
-  `,
-  );
-});
-
-test("decodeCallOnIntegrationArgs for Idle V4 lend should be equal to encoded data with encodeCallOnIntegrationArgs", () => {
-  const params = {
-    adapter: getAddress(IDLE_V4_ADAPTER),
-    selector: LEND_SELECTOR as Hex,
-    integrationData: toHex(
-      "000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000002000000000000000000000000fedc73464dfd156d30f6524654a5d56e766da0c3000000000000000000000000faf2c3db614e9d38fe05edc634848be7ff0542b90000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000003e8000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000033b2e3cce25d9e52486e3880000000000000000000000000000000000000000000000000000000000000000",
-    ),
-  };
-  const encoded = encodeCallOnIntegrationArgs(params);
-  const decoded = decodeCallOnIntegrationArgs(encoded);
-
-  expect(decoded).toEqual(params);
-});
-
-test("encodeCallOnIntegrationArgs for Idle V4 redeem should encode correctly", () => {
-  expect(
-    encodeCallOnIntegrationArgs({
-      adapter: IDLE_V4_ADAPTER,
-      selector: REDEEM_SELECTOR,
-      integrationData: "0x1230",
-    }),
-  ).toMatchInlineSnapshot(
-    '"0x000000000000000000000000474c8d4a0e53b7235c6f8fc27c9b6406a32dd0b1c29fa9dd00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000021230000000000000000000000000000000000000000000000000000000000000"',
-  );
-});
-
-test("decodeCallOnIntegrationArgs for Idle V4 redeem should decode correctly", () => {
-  expect(
-    decodeCallOnIntegrationArgs(
-      "0x000000000000000000000000474c8d4a0e53b7235c6f8fc27c9b6406a32dd0b1c29fa9dd00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000021230000000000000000000000000000000000000000000000000000000000000",
-    ),
-  ).toMatchInlineSnapshot(
-    {
-      adapter: IDLE_V4_ADAPTER,
-      selector: REDEEM_SELECTOR,
-      integrationData: "0x1230",
-    },
-    `
-    {
-      "adapter": "0x474c8d4A0e53B7235C6f8fC27c9B6406a32Dd0b1",
-      "integrationData": "0x1230",
-      "selector": "0xc29fa9dd",
-    }
-  `,
-  );
-});
-
-test("decodeCallOnIntegrationArgs for Idle V4 redeem should be equal to encoded data with encodeCallOnIntegrationArgs", () => {
-  const params = {
-    adapter: getAddress(IDLE_V4_ADAPTER),
-    selector: REDEEM_SELECTOR as Hex,
-    integrationData: toHex(
-      "000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000002000000000000000000000000fedc73464dfd156d30f6524654a5d56e766da0c3000000000000000000000000faf2c3db614e9d38fe05edc634848be7ff0542b90000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000003e8000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000033b2e3cce25d9e52486e3880000000000000000000000000000000000000000000000000000000000000000",
-    ),
-  };
-  const encoded = encodeCallOnIntegrationArgs(params);
-  const decoded = decodeCallOnIntegrationArgs(encoded);
-
-  expect(decoded).toEqual(params);
-});
-
-test("encodeCallOnIntegrationArgs for Idle V4 claim rewards should encode correctly", () => {
-  expect(
-    encodeCallOnIntegrationArgs({
-      adapter: IDLE_V4_ADAPTER,
-      selector: CLAIM_REWARDS_SELECTOR,
-      integrationData: "0x1230",
-    }),
-  ).toMatchInlineSnapshot(
-    '"0x000000000000000000000000474c8d4a0e53b7235c6f8fc27c9b6406a32dd0b1b9dfbacc00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000021230000000000000000000000000000000000000000000000000000000000000"',
-  );
-});
-
-test("decodeCallOnIntegrationArgs for Idle V4 claim rewards should decode correctly", () => {
-  expect(
-    decodeCallOnIntegrationArgs(
-      "0x000000000000000000000000474c8d4a0e53b7235c6f8fc27c9b6406a32dd0b1b9dfbacc00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000021230000000000000000000000000000000000000000000000000000000000000",
-    ),
-  ).toMatchInlineSnapshot(
-    {
-      adapter: IDLE_V4_ADAPTER,
-      selector: CLAIM_REWARDS_SELECTOR,
-      integrationData: "0x1230",
-    },
-    `
-    {
-      "adapter": "0x474c8d4A0e53B7235C6f8fC27c9B6406a32Dd0b1",
-      "integrationData": "0x1230",
-      "selector": "0xb9dfbacc",
-    }
-  `,
-  );
-});
-
-test("decodeCallOnIntegrationArgs for Idle V4 claim rewards should be equal to encoded data with encodeCallOnIntegrationArgs", () => {
-  const params = {
-    adapter: getAddress(IDLE_V4_ADAPTER),
-    selector: CLAIM_REWARDS_SELECTOR as Hex,
-    integrationData: toHex(
-      "000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000002000000000000000000000000fedc73464dfd156d30f6524654a5d56e766da0c3000000000000000000000000faf2c3db614e9d38fe05edc634848be7ff0542b90000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000003e8000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000033b2e3cce25d9e52486e3880000000000000000000000000000000000000000000000000000000000000000",
-    ),
-  };
-  const encoded = encodeCallOnIntegrationArgs(params);
-  const decoded = decodeCallOnIntegrationArgs(encoded);
-
-  expect(decoded).toEqual(params);
-});
