--- packages/sdk/src/extensions/integrations/callOnIntegration/uniswapV2Exchange.test.ts
@@ -1,52 +1,0 @@
-import { UNISWAP_V2_EXCHANGE_ADAPTER } from "../../../../tests/constants.js";
-import { TAKE_ORDER_SELECTOR } from "../../../constants/selectors.js";
-import { decodeCallOnIntegrationArgs, encodeCallOnIntegrationArgs } from "./callOnIntegration.js";
-import { getAddress, toHex } from "viem";
-import { expect, test } from "vitest";
-
-test("encodeCallOnIntegrationArgs for Uniswap V2 Exchange take order should encode correctly", () => {
-  expect(
-    encodeCallOnIntegrationArgs({
-      adapter: UNISWAP_V2_EXCHANGE_ADAPTER,
-      selector: TAKE_ORDER_SELECTOR,
-      integrationData: "0x1230",
-    }),
-  ).toMatchInlineSnapshot(
-    '"0x0000000000000000000000008c36435a653041bfd65515cc82502663c1ce6f0e03e38a2b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000021230000000000000000000000000000000000000000000000000000000000000"',
-  );
-});
-
-test("decodeCallOnIntegrationArgs for Uniswap V2 Exchange take order should decode correctly", () => {
-  expect(
-    decodeCallOnIntegrationArgs(
-      "0x0000000000000000000000008c36435a653041bfd65515cc82502663c1ce6f0e03e38a2b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000021230000000000000000000000000000000000000000000000000000000000000",
-    ),
-  ).toMatchInlineSnapshot(
-    {
-      adapter: UNISWAP_V2_EXCHANGE_ADAPTER,
-      selector: TAKE_ORDER_SELECTOR,
-      integrationData: "0x1230",
-    },
-    `
-    {
-      "adapter": "0x8c36435A653041BFd65515CC82502663C1ce6F0e",
-      "integrationData": "0x1230",
-      "selector": "0x03e38a2b",
-    }
-  `,
-  );
-});
-
-test("decodeCallOnIntegrationArgs for Uniswap V2 Exchange take order should be equal to encoded data with encodeCallOnIntegrationArgs", () => {
-  const params = {
-    adapter: getAddress(UNISWAP_V2_EXCHANGE_ADAPTER),
-    selector: TAKE_ORDER_SELECTOR,
-    integrationData: toHex(
-      "000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000002000000000000000000000000fedc73464dfd156d30f6524654a5d56e766da0c3000000000000000000000000faf2c3db614e9d38fe05edc634848be7ff0542b90000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000003e8000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000033b2e3cce25d9e52486e3880000000000000000000000000000000000000000000000000000000000000000",
-    ),
-  };
-  const encoded = encodeCallOnIntegrationArgs(params);
-  const decoded = decodeCallOnIntegrationArgs(encoded);
-
-  expect(decoded).toEqual(params);
-});
