+++ packages/sdk/src/extensions/integrations/instances/aaveV3.test.ts
@@ -1,78 +1,105 @@
-import { AAVE_V3_A_WETH } from "../../../../tests/constants.js";
+import { AAVE_V3_ADAPTER, AAVE_V3_A_WETH, ALICE, BOB, INTEGRATION_MANAGER, WETH } from "../../../../tests/constants.js";
+import { sendTestTransaction, testActions } from "../../../../tests/globals.js";
 import { toWei } from "../../../utils/conversion.js";
-import {
-  decodeAaveV3LendArgs,
-  decodeAaveV3RedeemArgs,
-  encodeAaveV3LendArgs,
-  encodeAaveV3RedeemArgs,
-} from "./aaveV3.js";
-import { getAddress } from "viem";
-import { expect, test } from "vitest";
+import { Integration } from "../integrationTypes.js";
+import { prepareUseIntegration } from "../prepareUseIntegration.js";
+import { test } from "vitest";
 
-test("decodeAaveV3LendArgs should be equal to encoded data with encodeAaveV3LendArgs", () => {
-  const params = {
-    aToken: getAddress(AAVE_V3_A_WETH),
-    depositAmount: toWei(100),
-  };
+test("prepare adapter trade for Aave V3 lend should work correctly", async () => {
+  const vaultOwner = ALICE;
+  const sharesBuyer = BOB;
 
-  const encoded = encodeAaveV3LendArgs(params);
-  const decoded = decodeAaveV3LendArgs(encoded);
+  const { comptrollerProxy, vaultProxy } = await testActions.createTestVault({
+    vaultOwner,
+    denominationAsset: WETH,
+  });
 
-  expect(decoded).toEqual(params);
-});
+  const depositAmount = toWei(250);
 
-test("encodeAaveV3LendArgs should encode correctly", () => {
-  expect(
-    encodeAaveV3LendArgs({
-      aToken: AAVE_V3_A_WETH,
-      depositAmount: toWei(100),
+  await testActions.buyShares({
+    comptrollerProxy,
+    sharesBuyer,
+    investmentAmount: depositAmount,
+  });
+
+  await sendTestTransaction({
+    ...prepareUseIntegration({
+      integrationManager: INTEGRATION_MANAGER,
+      integrationAdapter: AAVE_V3_ADAPTER,
+      callArgs: {
+        type: Integration.AaveV3Lend,
+        aToken: AAVE_V3_A_WETH,
+        depositAmount,
+      },
     }),
-  ).toMatchInlineSnapshot(
-    '"0x0000000000000000000000004d5f47fa6a74757f35c14fd3a6ef8e3c9bc514e80000000000000000000000000000000000000000000000056bc75e2d63100000"',
-  );
-});
+    account: vaultOwner,
+    address: comptrollerProxy,
+  });
 
-test("decodeAaveV3LendArgs should decode correctly", () => {
-  expect(
-    decodeAaveV3LendArgs(
-      "0x0000000000000000000000004d5f47fa6a74757f35c14fd3a6ef8e3c9bc514e80000000000000000000000000000000000000000000000056bc75e2d63100000",
-    ),
-  ).toEqual({
-    aToken: AAVE_V3_A_WETH,
-    depositAmount: toWei(100),
+  await testActions.assertBalanceOf({
+    token: AAVE_V3_A_WETH,
+    account: vaultProxy,
+    expected: depositAmount,
+    fuzziness: 100n,
   });
 });
 
-test("decodeAaveV3RedeemArgs should be equal to encoded data with encodeAaveV3RedeemArgs", () => {
-  const params = {
-    aToken: getAddress(AAVE_V3_A_WETH),
-    redeemAmount: toWei(100),
-  };
+test("prepare adapter trade for Aave V3 redeem should work correctly", async () => {
+  const vaultOwner = ALICE;
+  const sharesBuyer = BOB;
 
-  const encoded = encodeAaveV3RedeemArgs(params);
-  const decoded = decodeAaveV3RedeemArgs(encoded);
+  const { comptrollerProxy, vaultProxy } = await testActions.createTestVault({
+    vaultOwner,
+    denominationAsset: WETH,
+  });
 
-  expect(decoded).toEqual(params);
-});
+  const investmentAmount = toWei(250);
 
-test("encodeAaveV3RedeemArgs should encode correctly", () => {
-  expect(
-    encodeAaveV3RedeemArgs({
-      aToken: AAVE_V3_A_WETH,
-      redeemAmount: toWei(100),
+  await testActions.buyShares({
+    comptrollerProxy,
+    sharesBuyer,
+    investmentAmount: investmentAmount,
+  });
+
+  await sendTestTransaction({
+    ...prepareUseIntegration({
+      integrationManager: INTEGRATION_MANAGER,
+      integrationAdapter: AAVE_V3_ADAPTER,
+      callArgs: {
+        type: Integration.AaveV3Lend,
+        aToken: AAVE_V3_A_WETH,
+        depositAmount: investmentAmount,
+      },
     }),
-  ).toMatchInlineSnapshot(
-    '"0x0000000000000000000000004d5f47fa6a74757f35c14fd3a6ef8e3c9bc514e80000000000000000000000000000000000000000000000056bc75e2d63100000"',
-  );
-});
+    account: vaultOwner,
+    address: comptrollerProxy,
+  });
 
-test("decodeAaveV3RedeemArgs should decode correctly", () => {
-  expect(
-    decodeAaveV3RedeemArgs(
-      "0x0000000000000000000000004d5f47fa6a74757f35c14fd3a6ef8e3c9bc514e80000000000000000000000000000000000000000000000056bc75e2d63100000",
-    ),
-  ).toEqual({
-    aToken: AAVE_V3_A_WETH,
-    redeemAmount: toWei(100),
+  await testActions.assertBalanceOf({
+    token: AAVE_V3_A_WETH,
+    account: vaultProxy,
+    expected: investmentAmount,
+    fuzziness: 100n,
+  });
+
+  await sendTestTransaction({
+    ...prepareUseIntegration({
+      integrationManager: INTEGRATION_MANAGER,
+      integrationAdapter: AAVE_V3_ADAPTER,
+      callArgs: {
+        type: Integration.AaveV3Redeem,
+        aToken: AAVE_V3_A_WETH,
+        redeemAmount: investmentAmount,
+      },
+    }),
+    account: vaultOwner,
+    address: comptrollerProxy,
+  });
+
+  await testActions.assertBalanceOf({
+    token: WETH,
+    account: vaultProxy,
+    expected: investmentAmount,
+    fuzziness: 100n,
   });
 });
