+++ packages/sdk/src/extensions/integrations/instances/compoundV3.test.ts
@@ -1,111 +1,180 @@
-import { COMPOUND_V3_C_WETH } from "../../../../tests/constants.js";
-import { toWei } from "../../../utils/conversion.js";
+import { increaseTimeAndMine } from "../../../../tests/actions/increaseTimeAndMine.js";
 import {
-  decodeCompoundV3ClaimRewardsArgs,
-  decodeCompoundV3LendArgs,
-  decodeCompoundV3RedeemArgs,
-  encodeCompoundV3ClaimRewardsArgs,
-  encodeCompoundV3LendArgs,
-  encodeCompoundV3RedeemArgs,
-} from "./compoundV3.js";
-import { getAddress } from "viem";
+  ALICE,
+  BOB,
+  COMP,
+  COMPOUND_V3_ADAPTER,
+  COMPOUND_V3_C_WETH,
+  INTEGRATION_MANAGER,
+  WETH,
+} from "../../../../tests/constants.js";
+import { sendTestTransaction, testActions } from "../../../../tests/globals.js";
+import { toSeconds, toWei } from "../../../utils/conversion.js";
+import { Integration } from "../integrationTypes.js";
+import { prepareUseIntegration } from "../prepareUseIntegration.js";
 import { expect, test } from "vitest";
 
-test("decodeCompoundV3LendArgs should be equal to encoded data with encodeCompoundV3LendArgs", () => {
-  const params = {
-    cToken: getAddress(COMPOUND_V3_C_WETH),
-    depositAmount: toWei(100),
-  };
+test("prepare adapter trade for Compound V3 lend should work correctly", async () => {
+  const vaultOwner = ALICE;
+  const sharesBuyer = BOB;
 
-  const encoded = encodeCompoundV3LendArgs(params);
-  const decoded = decodeCompoundV3LendArgs(encoded);
+  const { comptrollerProxy, vaultProxy } = await testActions.createTestVault({
+    vaultOwner,
+    denominationAsset: WETH,
+  });
 
-  expect(decoded).toEqual(params);
-});
+  const depositAmount = toWei(250);
 
-test("encodeCompoundV3LendArgs should encode correctly", () => {
-  expect(
-    encodeCompoundV3LendArgs({
-      cToken: COMPOUND_V3_C_WETH,
-      depositAmount: toWei(100),
+  await testActions.buyShares({
+    comptrollerProxy,
+    sharesBuyer,
+    investmentAmount: depositAmount,
+  });
+
+  await sendTestTransaction({
+    ...prepareUseIntegration({
+      integrationManager: INTEGRATION_MANAGER,
+      integrationAdapter: COMPOUND_V3_ADAPTER,
+      callArgs: {
+        type: Integration.CompoundV3Lend,
+        cToken: COMPOUND_V3_C_WETH,
+        depositAmount,
+      },
     }),
-  ).toMatchInlineSnapshot(
-    '"0x000000000000000000000000a17581a9e3356d9a858b789d68b4d866e593ae940000000000000000000000000000000000000000000000056bc75e2d63100000"',
-  );
-});
+    account: vaultOwner,
+    address: comptrollerProxy,
+  });
 
-test("decodeCompoundV3LendArgs should decode correctly", () => {
-  expect(
-    decodeCompoundV3LendArgs(
-      "0x000000000000000000000000a17581a9e3356d9a858b789d68b4d866e593ae940000000000000000000000000000000000000000000000056bc75e2d63100000",
-    ),
-  ).toEqual({
-    cToken: COMPOUND_V3_C_WETH,
-    depositAmount: toWei(100),
+  await testActions.assertBalanceOf({
+    token: COMPOUND_V3_C_WETH,
+    account: vaultProxy,
+    expected: depositAmount,
+    fuzziness: 100n,
   });
 });
 
-test("decodeCompoundV3RedeemArgs should be equal to encoded data with encodeCompoundV3RedeemArgs", () => {
-  const params = {
-    cToken: getAddress(COMPOUND_V3_C_WETH),
-    redeemAmount: toWei(100),
-  };
+test("prepare adapter trade for Compound V3 redeem should work correctly", async () => {
+  const vaultOwner = ALICE;
+  const sharesBuyer = BOB;
 
-  const encoded = encodeCompoundV3RedeemArgs(params);
-  const decoded = decodeCompoundV3RedeemArgs(encoded);
+  const { comptrollerProxy, vaultProxy } = await testActions.createTestVault({
+    vaultOwner,
+    denominationAsset: WETH,
+  });
 
-  expect(decoded).toEqual(params);
-});
+  const investmentAmount = toWei(100);
 
-test("encodeCompoundV3RedeemArgs should encode correctly", () => {
-  expect(
-    encodeCompoundV3RedeemArgs({
-      cToken: COMPOUND_V3_C_WETH,
-      redeemAmount: toWei(100),
+  await testActions.buyShares({
+    comptrollerProxy,
+    sharesBuyer,
+    investmentAmount: investmentAmount,
+  });
+
+  await sendTestTransaction({
+    ...prepareUseIntegration({
+      integrationManager: INTEGRATION_MANAGER,
+      integrationAdapter: COMPOUND_V3_ADAPTER,
+      callArgs: {
+        type: Integration.CompoundV3Lend,
+        cToken: COMPOUND_V3_C_WETH,
+        depositAmount: investmentAmount,
+      },
     }),
-  ).toMatchInlineSnapshot(
-    '"0x000000000000000000000000a17581a9e3356d9a858b789d68b4d866e593ae940000000000000000000000000000000000000000000000056bc75e2d63100000"',
-  );
-});
+    account: vaultOwner,
+    address: comptrollerProxy,
+  });
 
-test("decodeCompoundV3RedeemArgs should decode correctly", () => {
-  expect(
-    decodeCompoundV3RedeemArgs(
-      "0x000000000000000000000000a17581a9e3356d9a858b789d68b4d866e593ae940000000000000000000000000000000000000000000000056bc75e2d63100000",
-    ),
-  ).toEqual({
-    cToken: COMPOUND_V3_C_WETH,
-    redeemAmount: toWei(100),
+  await testActions.assertBalanceOf({
+    token: COMPOUND_V3_C_WETH,
+    account: vaultProxy,
+    expected: investmentAmount,
+    fuzziness: 100n,
+  });
+
+  const redeemAmount = toWei(50);
+
+  await sendTestTransaction({
+    ...prepareUseIntegration({
+      integrationManager: INTEGRATION_MANAGER,
+      integrationAdapter: COMPOUND_V3_ADAPTER,
+      callArgs: {
+        type: Integration.CompoundV3Redeem,
+        cToken: COMPOUND_V3_C_WETH,
+        redeemAmount,
+      },
+    }),
+    account: vaultOwner,
+    address: comptrollerProxy,
+  });
+
+  await testActions.assertBalanceOf({
+    token: WETH,
+    account: vaultProxy,
+    expected: redeemAmount,
+    fuzziness: 100n,
   });
 });
 
-test("decodeCompoundV3ClaimRewardsArgs should be equal to encoded data with encodeCompoundV3ClaimRewardsArgs", () => {
-  const params = {
-    cTokens: [getAddress(COMPOUND_V3_C_WETH)],
-  };
+test("prepare adapter trade for Compound V3 claim rewards should work correctly", async () => {
+  const vaultOwner = ALICE;
+  const sharesBuyer = BOB;
 
-  const encoded = encodeCompoundV3ClaimRewardsArgs(params);
-  const decoded = decodeCompoundV3ClaimRewardsArgs(encoded);
-
-  expect(decoded).toEqual(params);
-});
-
-test("encodeCompoundV3ClaimRewardsArgs should encode correctly", () => {
-  expect(
-    encodeCompoundV3ClaimRewardsArgs({
-      cTokens: [COMPOUND_V3_C_WETH],
-    }),
-  ).toMatchInlineSnapshot(
-    '"0x00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000001000000000000000000000000a17581a9e3356d9a858b789d68b4d866e593ae94"',
-  );
-});
-
-test("decodeCompoundV3ClaimRewardsArgs should decode correctly", () => {
-  expect(
-    decodeCompoundV3ClaimRewardsArgs(
-      "0x00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000001000000000000000000000000a17581a9e3356d9a858b789d68b4d866e593ae94",
-    ),
-  ).toEqual({
-    cTokens: [COMPOUND_V3_C_WETH],
+  const { comptrollerProxy, vaultProxy } = await testActions.createTestVault({
+    vaultOwner,
+    denominationAsset: WETH,
   });
+
+  const investmentAmount = toWei(250);
+
+  await testActions.buyShares({
+    comptrollerProxy,
+    sharesBuyer,
+    investmentAmount: investmentAmount,
+  });
+
+  await sendTestTransaction({
+    ...prepareUseIntegration({
+      integrationManager: INTEGRATION_MANAGER,
+      integrationAdapter: COMPOUND_V3_ADAPTER,
+      callArgs: {
+        type: Integration.CompoundV3Lend,
+        cToken: COMPOUND_V3_C_WETH,
+        depositAmount: investmentAmount,
+      },
+    }),
+    account: vaultOwner,
+    address: comptrollerProxy,
+  });
+
+  await testActions.assertBalanceOf({
+    token: COMPOUND_V3_C_WETH,
+    account: vaultProxy,
+    expected: investmentAmount,
+    fuzziness: 100n,
+  });
+
+  await increaseTimeAndMine({
+    seconds: toSeconds({ hours: 1_000 }),
+    blocks: 1_000,
+  });
+
+  await sendTestTransaction({
+    ...prepareUseIntegration({
+      integrationManager: INTEGRATION_MANAGER,
+      integrationAdapter: COMPOUND_V3_ADAPTER,
+      callArgs: {
+        type: Integration.CompoundV3ClaimRewards,
+        cTokens: [COMPOUND_V3_C_WETH],
+      },
+    }),
+    account: vaultOwner,
+    address: comptrollerProxy,
+  });
+
+  const compBalance = await testActions.getBalanceOf({
+    token: COMP,
+    account: vaultProxy,
+  });
+
+  expect(compBalance).toBeGreaterThan(0n);
 });
