+++ packages/sdk/src/extensions/integrations/instances/idleV4.test.ts
@@ -1,110 +1,274 @@
-import { IDLE_V4_WETH } from "../../../../tests/constants.js";
-import { toWei } from "../../../utils/conversion.js";
+import { increaseTimeAndMine } from "../../../../tests/actions/increaseTimeAndMine.js";
 import {
-  decodeIdleV4ClaimRewardsArgs,
-  decodeIdleV4LendArgs,
-  decodeIdleV4RedeemArgs,
-  encodeIdleV4ClaimRewardsArgs,
-  encodeIdleV4LendArgs,
-  encodeIdleV4RedeemArgs,
-} from "./idleV4.js";
+  ALICE,
+  BOB,
+  IDLE,
+  IDLE_V4_ADAPTER,
+  IDLE_V4_PRICE_FEED,
+  IDLE_V4_WETH,
+  INTEGRATION_MANAGER,
+  WETH,
+} from "../../../../tests/constants.js";
+import { publicClient, sendTestTransaction, testActions, testClient } from "../../../../tests/globals.js";
+import { toSeconds, toWei } from "../../../utils/conversion.js";
+import { multiplyByRate } from "../../../utils/rates.js";
+import { multiplyBySlippage } from "../../../utils/slippage.js";
+import { prepareFunctionParams } from "../../../utils/viem.js";
+import { IIdlePriceFeed } from "../../abis/index.js";
+import { Integration } from "../integrationTypes.js";
+import { prepareUseIntegration } from "../prepareUseIntegration.js";
+import { type Address, getAbiItem, parseUnits, zeroAddress } from "viem";
 import { expect, test } from "vitest";
 
-test("decodeIdleV4LendArgs should be equal to encoded data with encodeIdleV4LendArgs", () => {
-  const params = {
-    idleToken: IDLE_V4_WETH,
-    depositAmount: toWei(100),
-    minIncomingIdleTokenAmount: toWei(50),
-  } as const;
-
-  const encoded = encodeIdleV4LendArgs(params);
-  const decoded = decodeIdleV4LendArgs(encoded);
-
-  expect(decoded).toEqual(params);
-});
-
-test("encodeIdleV4LendArgs should encode correctly", () => {
-  expect(
-    encodeIdleV4LendArgs({
-      idleToken: IDLE_V4_WETH,
-      depositAmount: toWei(100),
-      minIncomingIdleTokenAmount: toWei(50),
+async function getIdlePoolTokenRate(idlePoolToken: Address) {
+  const { result } = await publicClient.simulateContract({
+    ...prepareFunctionParams({
+      abi: getAbiItem({ abi: IIdlePriceFeed, name: "calcUnderlyingValues" }),
+      args: [idlePoolToken, parseUnits("1", 18)],
     }),
-  ).toMatchInlineSnapshot(
-    '"0x000000000000000000000000c8e6ca6e96a326dc448307a5fde90a0b21fd7f800000000000000000000000000000000000000000000000056bc75e2d63100000000000000000000000000000000000000000000000000002b5e3af16b1880000"',
-  );
-});
+    address: IDLE_V4_PRICE_FEED,
+    account: zeroAddress,
+  });
 
-test("decodeIdleV4LendArgs should decode correctly", () => {
-  expect(
-    decodeIdleV4LendArgs(
-      "0x000000000000000000000000c8e6ca6e96a326dc448307a5fde90a0b21fd7f800000000000000000000000000000000000000000000000056bc75e2d63100000000000000000000000000000000000000000000000000002b5e3af16b1880000",
-    ),
-  ).toEqual({
-    idleToken: IDLE_V4_WETH,
-    depositAmount: toWei(100),
-    minIncomingIdleTokenAmount: toWei(50),
+  const [_, underlyingAmounts] = result;
+
+  const rate = underlyingAmounts[0];
+
+  if (rate === undefined) {
+    throw new Error("Rate is undefined");
+  }
+
+  return rate;
+}
+
+test("prepare adapter trade for Idle V4 lend should work correctly", async () => {
+  const vaultOwner = ALICE;
+  const sharesBuyer = BOB;
+
+  const { comptrollerProxy, vaultProxy } = await testActions.createTestVault({
+    vaultOwner,
+    denominationAsset: WETH,
+  });
+
+  const depositAmount = toWei(250);
+
+  await testActions.buyShares({
+    comptrollerProxy,
+    sharesBuyer,
+    investmentAmount: depositAmount,
+  });
+
+  const poolTokenRate = await getIdlePoolTokenRate(IDLE_V4_WETH);
+
+  const slippage = 1n;
+
+  const minIncomingIdleTokenAmount = multiplyByRate({
+    inverse: true,
+    rate: poolTokenRate,
+    rateDecimals: 18,
+    value: depositAmount,
+  });
+
+  const minIncomingIdleTokenAmountWithSlippage = multiplyBySlippage({
+    amount: minIncomingIdleTokenAmount,
+    slippage,
+  });
+
+  await sendTestTransaction({
+    ...prepareUseIntegration({
+      integrationManager: INTEGRATION_MANAGER,
+      integrationAdapter: IDLE_V4_ADAPTER,
+      callArgs: {
+        type: Integration.IdleV4Lend,
+        idleToken: IDLE_V4_WETH,
+        depositAmount,
+        minIncomingIdleTokenAmount: minIncomingIdleTokenAmountWithSlippage,
+      },
+    }),
+    account: vaultOwner,
+    address: comptrollerProxy,
+  });
+
+  await testActions.assertBalanceOf({
+    token: IDLE_V4_WETH,
+    account: vaultProxy,
+    expected: minIncomingIdleTokenAmount,
+    fuzziness: minIncomingIdleTokenAmount - minIncomingIdleTokenAmountWithSlippage,
   });
 });
 
-test("decodeIdleV4RedeemArgs should be equal to encoded data with encodeIdleV4RedeemArgs", () => {
-  const params = {
-    idleToken: IDLE_V4_WETH,
-    outgoingIdleTokenAmount: toWei(100),
-    minIncomingUnderlyingAmount: toWei(50),
-  } as const;
+test("prepare adapter trade for Idle V4 redeem should work correctly", async () => {
+  const vaultOwner = ALICE;
+  const sharesBuyer = BOB;
 
-  const encoded = encodeIdleV4RedeemArgs(params);
-  const decoded = decodeIdleV4RedeemArgs(encoded);
+  const { comptrollerProxy, vaultProxy } = await testActions.createTestVault({
+    vaultOwner,
+    denominationAsset: WETH,
+  });
 
-  expect(decoded).toEqual(params);
-});
+  const investmentAmount = toWei(250);
 
-test("encodeIdleV4RedeemArgs should encode correctly", () => {
-  expect(
-    encodeIdleV4RedeemArgs({
-      idleToken: IDLE_V4_WETH,
-      outgoingIdleTokenAmount: toWei(100),
-      minIncomingUnderlyingAmount: toWei(50),
+  await testActions.buyShares({
+    comptrollerProxy,
+    sharesBuyer,
+    investmentAmount: investmentAmount,
+  });
+
+  const poolTokenRateBeforeLend = await getIdlePoolTokenRate(IDLE_V4_WETH);
+
+  const slippage = 1n;
+
+  const minIncomingIdleTokenAmount = multiplyByRate({
+    inverse: true,
+    rate: poolTokenRateBeforeLend,
+    rateDecimals: 18,
+    value: investmentAmount,
+  });
+
+  const minIncomingIdleTokenAmountWithSlippage = multiplyBySlippage({
+    amount: minIncomingIdleTokenAmount,
+    slippage,
+  });
+
+  await sendTestTransaction({
+    ...prepareUseIntegration({
+      integrationManager: INTEGRATION_MANAGER,
+      integrationAdapter: IDLE_V4_ADAPTER,
+      callArgs: {
+        type: Integration.IdleV4Lend,
+        idleToken: IDLE_V4_WETH,
+        depositAmount: investmentAmount,
+        minIncomingIdleTokenAmount: minIncomingIdleTokenAmountWithSlippage,
+      },
     }),
-  ).toMatchInlineSnapshot(
-    '"0x000000000000000000000000c8e6ca6e96a326dc448307a5fde90a0b21fd7f800000000000000000000000000000000000000000000000056bc75e2d63100000000000000000000000000000000000000000000000000002b5e3af16b1880000"',
-  );
-});
+    account: vaultOwner,
+    address: comptrollerProxy,
+  });
 
-test("decodeIdleV4RedeemArgs should decode correctly", () => {
-  expect(
-    decodeIdleV4RedeemArgs(
-      "0x000000000000000000000000c8e6ca6e96a326dc448307a5fde90a0b21fd7f800000000000000000000000000000000000000000000000056bc75e2d63100000000000000000000000000000000000000000000000000002b5e3af16b1880000",
-    ),
-  ).toEqual({
-    idleToken: IDLE_V4_WETH,
-    outgoingIdleTokenAmount: toWei(100),
-    minIncomingUnderlyingAmount: toWei(50),
+  await testActions.assertBalanceOf({
+    token: IDLE_V4_WETH,
+    account: vaultProxy,
+    expected: minIncomingIdleTokenAmount,
+    fuzziness: minIncomingIdleTokenAmount - minIncomingIdleTokenAmountWithSlippage,
+  });
+
+  const poolTokenRateBeforeRedeem = await getIdlePoolTokenRate(IDLE_V4_WETH);
+
+  const minIncomingUnderlyingAmount = multiplyByRate({
+    inverse: false,
+    rate: poolTokenRateBeforeRedeem,
+    rateDecimals: 18,
+    value: minIncomingIdleTokenAmountWithSlippage,
+  });
+
+  const minIncomingUnderlyingAmountWithSlippage = multiplyBySlippage({
+    amount: minIncomingUnderlyingAmount,
+    slippage,
+  });
+
+  // wait at least one block so reentrancy guard is not triggered
+  await testClient.mine({ blocks: 1 });
+
+  await sendTestTransaction({
+    ...prepareUseIntegration({
+      integrationManager: INTEGRATION_MANAGER,
+      integrationAdapter: IDLE_V4_ADAPTER,
+      callArgs: {
+        type: Integration.IdleV4Redeem,
+        idleToken: IDLE_V4_WETH,
+        outgoingIdleTokenAmount: minIncomingIdleTokenAmountWithSlippage,
+        minIncomingUnderlyingAmount: minIncomingUnderlyingAmountWithSlippage,
+      },
+    }),
+    account: vaultOwner,
+    address: comptrollerProxy,
+  });
+
+  await testActions.assertBalanceOf({
+    token: WETH,
+    account: vaultProxy,
+    expected: minIncomingUnderlyingAmount,
+    fuzziness: minIncomingUnderlyingAmount - minIncomingUnderlyingAmountWithSlippage,
   });
 });
 
-test("decodeIdleV4ClaimRewardsArgs should be equal to encoded data with encodeIdleV4ClaimRewardsArgs", () => {
-  const params = {
-    idleToken: IDLE_V4_WETH,
-  } as const;
+test("prepare adapter trade for Idle V4 claim rewards should work correctly", async () => {
+  const vaultOwner = ALICE;
+  const sharesBuyer = BOB;
 
-  const encoded = encodeIdleV4ClaimRewardsArgs(params);
-  const decoded = decodeIdleV4ClaimRewardsArgs(encoded);
-
-  expect(decoded).toEqual(params);
-});
-
-test("encodeIdleV4ClaimRewardsArgs should encode correctly", () => {
-  expect(
-    encodeIdleV4ClaimRewardsArgs({
-      idleToken: IDLE_V4_WETH,
-    }),
-  ).toMatchInlineSnapshot('"0x000000000000000000000000c8e6ca6e96a326dc448307a5fde90a0b21fd7f80"');
-});
-
-test("decodeIdleV4ClaimRewardsArgs should decode correctly", () => {
-  expect(decodeIdleV4ClaimRewardsArgs("0x000000000000000000000000c8e6ca6e96a326dc448307a5fde90a0b21fd7f80")).toEqual({
-    idleToken: IDLE_V4_WETH,
+  const { comptrollerProxy, vaultProxy } = await testActions.createTestVault({
+    vaultOwner,
+    denominationAsset: WETH,
   });
+
+  const investmentAmount = toWei(250);
+
+  await testActions.buyShares({
+    comptrollerProxy,
+    sharesBuyer,
+    investmentAmount: investmentAmount,
+  });
+
+  const poolTokenRateBeforeLend = await getIdlePoolTokenRate(IDLE_V4_WETH);
+
+  const slippage = 1n;
+
+  const minIncomingIdleTokenAmount = multiplyByRate({
+    inverse: true,
+    rate: poolTokenRateBeforeLend,
+    rateDecimals: 18,
+    value: investmentAmount,
+  });
+
+  const minIncomingIdleTokenAmountWithSlippage = multiplyBySlippage({
+    amount: minIncomingIdleTokenAmount,
+    slippage,
+  });
+
+  await sendTestTransaction({
+    ...prepareUseIntegration({
+      integrationManager: INTEGRATION_MANAGER,
+      integrationAdapter: IDLE_V4_ADAPTER,
+      callArgs: {
+        type: Integration.IdleV4Lend,
+        idleToken: IDLE_V4_WETH,
+        depositAmount: investmentAmount,
+        minIncomingIdleTokenAmount: minIncomingIdleTokenAmountWithSlippage,
+      },
+    }),
+    account: vaultOwner,
+    address: comptrollerProxy,
+  });
+
+  await testActions.assertBalanceOf({
+    token: IDLE_V4_WETH,
+    account: vaultProxy,
+    expected: minIncomingIdleTokenAmount,
+    fuzziness: minIncomingIdleTokenAmount - minIncomingIdleTokenAmountWithSlippage,
+  });
+
+  await increaseTimeAndMine({
+    seconds: toSeconds({ hours: 1_000 }),
+    blocks: 1_000,
+  });
+
+  await sendTestTransaction({
+    ...prepareUseIntegration({
+      integrationManager: INTEGRATION_MANAGER,
+      integrationAdapter: IDLE_V4_ADAPTER,
+      callArgs: {
+        type: Integration.IdleV4ClaimRewards,
+        idleToken: IDLE_V4_WETH,
+      },
+    }),
+    account: vaultOwner,
+    address: comptrollerProxy,
+  });
+
+  const idleBalance = await testActions.getBalanceOf({
+    token: IDLE,
+    account: vaultProxy,
+  });
+
+  expect(idleBalance).toBeGreaterThan(0n);
 });
