--- packages/sdk/src/extensions/integrations/prepareUseIntegration/aaveV2.test.ts
@@ -1,205 +1,0 @@
-import { AAVE_V2_ADAPTER, AAVE_V2_A_WETH, ALICE, BOB, INTEGRATION_MANAGER, WETH } from "../../../../tests/constants.js";
-import { sendTestTransaction, testActions } from "../../../../tests/globals.js";
-import { toWei } from "../../../utils/conversion.js";
-import { Integration } from "../integrationTypes.js";
-import { prepareUseIntegration } from "./prepareUseIntegration.js";
-import { expect, test } from "vitest";
-
-test("prepare adapter trade for Aave V2 lend should work correctly", async () => {
-  const vaultOwner = ALICE;
-  const sharesBuyer = BOB;
-
-  const { comptrollerProxy, vaultProxy } = await testActions.createTestVault({
-    vaultOwner,
-    denominationAsset: WETH,
-  });
-
-  const depositAmount = toWei(250);
-
-  await testActions.buyShares({
-    comptrollerProxy,
-    sharesBuyer,
-    investmentAmount: depositAmount,
-  });
-
-  await sendTestTransaction({
-    ...prepareUseIntegration({
-      integrationManager: INTEGRATION_MANAGER,
-      integrationAdapter: AAVE_V2_ADAPTER,
-      callArgs: {
-        type: Integration.AaveV2Lend,
-        aToken: AAVE_V2_A_WETH,
-        depositAmount,
-      },
-    }),
-    account: vaultOwner,
-    address: comptrollerProxy,
-  });
-
-  await testActions.assertBalanceOf({
-    token: AAVE_V2_A_WETH,
-    account: vaultProxy,
-    expected: depositAmount,
-    fuzziness: 100n,
-  });
-});
-
-test("prepareUseIntegration for Aave V2 lend should be equal to encoded data with encodeCallArgsForAaveV2Lend", () => {
-  expect(
-    prepareUseIntegration({
-      integrationManager: INTEGRATION_MANAGER,
-      integrationAdapter: AAVE_V2_ADAPTER,
-      callArgs: {
-        type: Integration.AaveV2Lend,
-        aToken: AAVE_V2_A_WETH,
-        depositAmount: toWei(100),
-      },
-    }),
-  ).toMatchInlineSnapshot(
-    `
-    {
-      "abi": [
-        {
-          "inputs": [
-            {
-              "internalType": "address",
-              "name": "_extension",
-              "type": "address",
-            },
-            {
-              "internalType": "uint256",
-              "name": "_actionId",
-              "type": "uint256",
-            },
-            {
-              "internalType": "bytes",
-              "name": "_callArgs",
-              "type": "bytes",
-            },
-          ],
-          "name": "callOnExtension",
-          "outputs": [],
-          "stateMutability": "nonpayable",
-          "type": "function",
-        },
-      ],
-      "args": [
-        "0x31329024f1a3E4a4B3336E0b1DfA74CC3FEc633e",
-        0n,
-        "0x000000000000000000000000ece6b376af7c9273cebaf6528565c47ea2cb8a4c099f75150000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000000040000000000000000000000000030ba81f1c18d280636f32af80b9aad02cf0854e0000000000000000000000000000000000000000000000056bc75e2d63100000",
-      ],
-      "functionName": "callOnExtension",
-    }
-  `,
-  );
-});
-
-test("prepare adapter trade for Aave V2 redeem should work correctly", async () => {
-  const vaultOwner = ALICE;
-  const sharesBuyer = BOB;
-
-  const { comptrollerProxy, vaultProxy } = await testActions.createTestVault({
-    vaultOwner,
-    denominationAsset: WETH,
-  });
-
-  const investmentAmount = toWei(250);
-
-  await testActions.buyShares({
-    comptrollerProxy,
-    sharesBuyer,
-    investmentAmount: investmentAmount,
-  });
-
-  await sendTestTransaction({
-    ...prepareUseIntegration({
-      integrationManager: INTEGRATION_MANAGER,
-      integrationAdapter: AAVE_V2_ADAPTER,
-      callArgs: {
-        type: Integration.AaveV2Lend,
-        aToken: AAVE_V2_A_WETH,
-        depositAmount: investmentAmount,
-      },
-    }),
-    account: vaultOwner,
-    address: comptrollerProxy,
-  });
-
-  await testActions.assertBalanceOf({
-    token: AAVE_V2_A_WETH,
-    account: vaultProxy,
-    expected: investmentAmount,
-    fuzziness: 100n,
-  });
-
-  await sendTestTransaction({
-    ...prepareUseIntegration({
-      integrationManager: INTEGRATION_MANAGER,
-      integrationAdapter: AAVE_V2_ADAPTER,
-      callArgs: {
-        type: Integration.AaveV2Redeem,
-        aToken: AAVE_V2_A_WETH,
-        redeemAmount: investmentAmount,
-      },
-    }),
-    account: vaultOwner,
-    address: comptrollerProxy,
-  });
-
-  await testActions.assertBalanceOf({
-    token: WETH,
-    account: vaultProxy,
-    expected: investmentAmount,
-    fuzziness: 100n,
-  });
-});
-
-test("prepareUseIntegration for Aave V2 redeem should be equal to encoded data with encodeCallArgsForAaveV2Redeem", () => {
-  expect(
-    prepareUseIntegration({
-      integrationManager: INTEGRATION_MANAGER,
-      integrationAdapter: AAVE_V2_ADAPTER,
-      callArgs: {
-        type: Integration.AaveV2Redeem,
-        aToken: AAVE_V2_A_WETH,
-        redeemAmount: toWei(100),
-      },
-    }),
-  ).toMatchInlineSnapshot(
-    `
-    {
-      "abi": [
-        {
-          "inputs": [
-            {
-              "internalType": "address",
-              "name": "_extension",
-              "type": "address",
-            },
-            {
-              "internalType": "uint256",
-              "name": "_actionId",
-              "type": "uint256",
-            },
-            {
-              "internalType": "bytes",
-              "name": "_callArgs",
-              "type": "bytes",
-            },
-          ],
-          "name": "callOnExtension",
-          "outputs": [],
-          "stateMutability": "nonpayable",
-          "type": "function",
-        },
-      ],
-      "args": [
-        "0x31329024f1a3E4a4B3336E0b1DfA74CC3FEc633e",
-        0n,
-        "0x000000000000000000000000ece6b376af7c9273cebaf6528565c47ea2cb8a4cc29fa9dd0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000000040000000000000000000000000030ba81f1c18d280636f32af80b9aad02cf0854e0000000000000000000000000000000000000000000000056bc75e2d63100000",
-      ],
-      "functionName": "callOnExtension",
-    }
-  `,
-  );
-});
